search: move TimeoutDuration to job logic
